requirements.txt
requests
pyyaml
crayons